refactor done

+++ libft/srcs/ft_strjoin.c
@@ -6,7 +6,7 @@
 /*   By: yeepark <yeepark@student.42seoul.k>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/11 14:50:26 by yeepark           #+#    #+#             */
-/*   Updated: 2022/07/12 16:36:06 by yeepark          ###   ########.fr       */
+/*   Updated: 2022/07/17 19:04:57 by yeepark          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strjoin(char const *s1, char const *s2)
 {
-	unsigned int	size;
-	char			*str;
+	size_t	size;
+	char	*str;
 
 	size = ft_strlen(s1) + ft_strlen(s2);
 	str = malloc(sizeof(char) * (size + 1));
